Time: 258 ms (67.67%), Space: 76.1 MB (85.94%) - LeetHub

+++ 1830-count-good-meals/1830-count-good-meals.cpp
@@ -16,7 +16,7 @@ public:
                     }
                     else
                     {
-                        ans+=(long long)(m[k-nums[i]]-1);
+                        ans+=(m[k-nums[i]]-1);
                     }
                 }
             }
